[MDH-50] refactor : 락 테스트 클래스 분리
