+++ .idea/modules.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/devtable.iml" filepath="$PROJECT_DIR$/.idea/modules/devtable.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/devtable.main.iml" filepath="$PROJECT_DIR$/.idea/modules/devtable.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/devtable.test.iml" filepath="$PROJECT_DIR$/.idea/modules/devtable.test.iml" />
+    </modules>
+  </component>
+</project>
